Reorganise codebase using classes
